return weighted average

+++ 0x04-python-more_data_structures/100-weight_average.py
@@ -11,4 +11,4 @@ def weight_average(my_list=[]):
 
         summation = summation + mul
 
-    return summation/divident
+    return float(summation/divident)
